Moved all test class to test folder
