+++ src/test/java/LisaTest/WithinTest.java
@@ -1,20 +1,23 @@
-package task;
+package LisaTest;
 
 import enumerations.PriorityEnum;
 import enumerations.SymbolEnum;
+import task.Task;
+import task.Within;
 
 import org.junit.Test;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 
 /**
- * This class tests the ToDo class.
+ * This class tests the Within class.
  *
  * @author Aloysius Wong
- * @Version 1.0
+ * @version 1.0
  * @since 08-11-2020
  */
-public class TodoTest {
+public class WithinTest {
 
     /**
      * Tests:
@@ -23,15 +26,23 @@ public class TodoTest {
      * 3) getSymbol
      * 4) getDone
      * 5) getPriority
+     * 6) Start LocalDateTime
+     * 7) End LocalDateTime
      */
     @Test
     public void constructorTest() {
-        Task test = new ToDo("test case");
-        assertEquals("test case", test.getDescription());
-        assertEquals(SymbolEnum.T, test.getSymbol());
+        Task test = new Within("tEsT cAsE", "10/10/2020 10:10", "23/8/2025 03:20");
+        assertEquals("tEsT cAsE", test.getDescription());
+        assertEquals(SymbolEnum.W, test.getSymbol());
         assertEquals("[✗]", test.getDone());
         assertFalse(test.isDone);
         assertEquals(PriorityEnum.NA, test.getPriority());
+        assertEquals(10, test.getStart().getDayOfMonth());
+        assertEquals(10, test.getStart().getMonthValue());
+        assertEquals(2020, test.getStart().getYear());
+        assertEquals(23, test.getEnd().getDayOfMonth());
+        assertEquals(8, test.getEnd().getMonthValue());
+        assertEquals(2025, test.getEnd().getYear());
     }
 
     /**
@@ -40,7 +51,7 @@ public class TodoTest {
      */
     @Test
     public void setDoneTest() {
-        Task test = new ToDo("test case");
+        Task test = new Within("tEsT cAsE", "10/10/2020 10:10", "23/8/2025 03:20");
         test.setDone();
         assertEquals("[✓]", test.getDone());
     }
@@ -51,7 +62,7 @@ public class TodoTest {
      */
     @Test
     public void setPriorityTest() {
-        Task test = new ToDo("test case");
+        Task test = new Within("tEsT cAsE", "10/10/2020 10:10", "23/8/2025 03:20");
         test.setPriority("hiGh");
         assertEquals(PriorityEnum.HIGH, test.getPriority());
         test.setPriority("MEDIUM");
